+++ packages/python/qsu/misc/__init__.py
@@ -2,10 +2,12 @@ from .debounce import debounce
 from .funcTimes import funcTimes
 from .logBox import logBox
 from .sleep import sleep
+from .throttle import throttle
 
 __all__ = [
 	'debounce',
 	'funcTimes',
 	'logBox',
 	'sleep',
+	'throttle',
 ]